+++ base-entities/src/main/java/de/alpharogroup/db/entity/traceable/TraceableEntity.java
@@ -27,14 +27,15 @@ import javax.persistence.MappedSuperclass;
 import de.alpharogroup.db.entity.BaseEntity;
 import de.alpharogroup.db.entity.create.Creation;
 import de.alpharogroup.db.entity.delete.Deletion;
+import de.alpharogroup.db.entity.modify.LastModification;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 /**
- * The entity class {@link TraceableEntity} is keeping the information for the deletion and for the
- * creation of an entity. <br>
+ * The entity class {@link TraceableEntity} is keeping the information for the creation, last
+ * modification and for the deletion of an entity. <br>
  *
  * @param <PK>
  *            the generic type of the id
@@ -42,6 +43,7 @@ import lombok.Setter;
  *            the generic type of the user or account
  * @see Creation
  * @see Deletion
+ * @see LastModification
  */
 @MappedSuperclass
 @Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
@@ -56,17 +58,22 @@ public abstract class TraceableEntity<PK extends Serializable, U> extends BaseEn
 	/** The Constant serialVersionUID. */
 	private static final long serialVersionUID = 1L;
 
-	/** The date and time when the entity that owns this entity was deleted. */
-	private LocalDateTime deleted;
-
-	/** The user or account that deleted the entity that owns this entity. */
-	private U deletedBy;
-
-
 	/** The date and time when the entity that owns this entity was created. */
 	private LocalDateTime created;
 
 	/** The user or account that created the entity that owns this entity. */
 	private U createdBy;
+
+	/** The date and time when the entity that owns this entity was modified. */
+	private LocalDateTime lastModified;
+
+	/** The user or account that modified the entity that owns this entity. */
+	private U lastModifiedBy;
+
+	/** The date and time when the entity that owns this entity was deleted. */
+	private LocalDateTime deleted;
+
+	/** The user or account that deleted the entity that owns this entity. */
+	private U deletedBy;
 
 }