+++ business-api/src/main/java/de/alpharogroup/db/service/api/BusinessService.java
@@ -18,6 +18,7 @@ package de.alpharogroup.db.service.api;
 import java.io.Serializable;
 import java.util.List;
 
+import javax.persistence.EntityManager;
 import javax.persistence.Query;
 
 import de.alpharogroup.db.entity.BaseEntity;
@@ -45,7 +46,7 @@ public interface BusinessService<T extends BaseEntity<PK>, PK extends Serializab
 	 *            a native SQL query string
 	 * @return the new {@link Query} instance
 	 */
-	public Query createNativeQuery(String sqlString);
+	Query createNativeQuery(String sqlString);
 
 	/**
 	 * Factory method for create a new instance of {@link Query} from the given sql string and
@@ -60,9 +61,7 @@ public interface BusinessService<T extends BaseEntity<PK>, PK extends Serializab
 	 * 
 	 * @return the new {@link Query} instance
 	 */
-	public Query createNativeQuery(String sqlString,
-		@SuppressWarnings("rawtypes") Class resultClass);
-
+	Query createNativeQuery(String sqlString, @SuppressWarnings("rawtypes") Class resultClass);
 
 	/**
 	 * Factory method for create a new instance of {@link Query} from the given sql string and
@@ -76,8 +75,8 @@ public interface BusinessService<T extends BaseEntity<PK>, PK extends Serializab
 	 *            the name of the result set mapping
 	 * @return the new {@link Query} instance
 	 */
-	public Query createNativeQuery(String sqlString, String resultSetMapping);
-	
+	Query createNativeQuery(String sqlString, String resultSetMapping);
+
 	/**
 	 * Delete all persistent objects in the given list.
 	 * 
@@ -135,6 +134,13 @@ public interface BusinessService<T extends BaseEntity<PK>, PK extends Serializab
 	 * @return An object of type T
 	 */
 	T get(PK id);
+
+	/**
+	 * Gets the entity manager.
+	 *
+	 * @return the entity manager
+	 */
+	EntityManager getEntityManager();
 
 	/**
 	 * Retrieve a persisted object with a given id from the database.